added cleanup to livestream script

+++ livestream_rip.sh
@@ -2,12 +2,12 @@
 # rips AZleg livestream in real-time, storing output in 20 minute chunks
 # version 1
 
-read -p "Target (url): " targ
-read -p "choose filename prefix: " outNAME
-targURL=$(curl -L $targ | tr "\'" "\n" |
-	grep "\Khttp.*?media.*?m3u?8" -Poz | tr -d '\0') &&
-	ffmpeg -i $targURL -c copy -segment_time 00:20:00 -f -reset_timestamps 1 \
-	segment $outNAME%03d.mp4
+# read -p "Target (url): " targ
+# read -p "choose filename prefix: " outNAME
+# targURL=$(curl -L $targ | tr "\'" "\n" |
+#	grep "\Khttp.*?media.*?m3u?8" -Poz | tr -d '\0') &&
+#	ffmpeg -i $targURL -c copy -segment_time 00:20:00 -f # -reset_timestamps 1 \
+#	segment $outNAME%03d.mp4
 
 # version 2
 read -p "Target (url): " targ
@@ -20,11 +20,14 @@ targURL=$(curl -L $targ | tr "\'" "\n" |
 	ffmpeg -i $targURL -c copy -segment_time $timestamp -reset_timestamps 1 \
 	-f segment $outNAME%03d.mp4
 
-# clean up: since break is partial, remove last segment
-rm $(ls $outNAME* | sort | tail -n 1)
+read -p "delete last segment? [y/n]: " opt
+if [[ $opt == "y" ]]; then
+    # clean up: since break is partial, remove last segment
+    rm $(ls $outNAME* | sort | tail -n 1);
+fi
 
 # merge segments into single file and create audio-only file for easy transcription
-echo $(ls $outNAME*) | sed 's/ /\n/g' |  sed 's/^/file /g' > temp &&
-ffmpeg -f concat -i temp -c copy "${outNAME}_all.mp4" && rm temp &&
+echo $(ls $outNAME*) | sed 's/ /\n/g' |  sed 's/^/file /g' > temp
+ffmpeg -f concat -i temp -c copy "${outNAME}_all.mp4" && rm temp
 ffmpeg -i "${outNAME}_all.mp4" -vn -ac 2 -b:a 192k "${outNAME}_all.mp3"
